move giftedCHat types
